% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_univariate_autocorr.R
\name{check_univariate_autocorr}
\alias{check_univariate_autocorr}
\title{Check for autocorrelation in a univariate time series}
\usage{
check_univariate_autocorr(
  serie,
  serie_name,
  test_type = c("Ljung-Box", "Box-Pierce"),
  n_lag = c(1, 20),
  print_tests = TRUE,
  return_output = TRUE,
  path_output = NULL
)
}
\arguments{
\item{serie}{Numerical vector containing the time series.}

\item{serie_name}{Character indicating the name of the serie.}

\item{test_type}{Character indicating the test to perform. Either "Ljung-Box"
(the default) or "Box-Pierce".}

\item{n_lag}{Numeric (vector) indicating the number of lags to be used for the second
test. If it is a vector, then the test will be done for each lag.}

\item{print_tests}{Logical indicating if the results of the tests should be
printed out. True by default.}

\item{return_output}{Logical indicating if the graph and the result's tibble
should be returned. TRUE by default.}

\item{path_output}{NULL or character indicating the path where to save
the graph. If NULL (the default), graph will not be saved.}
}
\value{
A list with 2 elements :
\itemize{
\item "graph" : the ACF and PACF graph
\item "df_res" : the tibble containing the tests p.values
}
}
\description{
Create ACF and PACF graphs for a univariate serie with
a confidence intervall of 5\%. Also
perform a Ljung-Box or Box-Pierce test for 1 and \code{n_lag}. The p.values
of these tests can be returned in a tibble.

For more informations on statistical tests see :
\code{\link[stats]{Box.test}}(). The null hypothesis is the abscence of
autocorrelation for a given number of lags.
}
\examples{
set.seed(123)  # Pour assurer la reproductibilité

# Normal independant distribution
normal_serie = stats::rnorm(1000, mean = 3, sd = 7)

# Create an AR1 process
n <- 1000       
phi <- 0.7      
epsilon <- stats::rnorm(n)  

serie_autocorr <- numeric(n)
serie_autocorr[1] <- epsilon[1] 

for (t in 2:n) {
  serie_autocorr[t] <- phi * serie_autocorr[t-1] + epsilon[t]
}

# Test the autocorrelation for one serie with a Ljung-Box test
check_univariate_autocorr(serie_autocorr, "AR(1)")

# Test the autocorrelation for the 2 series with the Box-Pierce test
# Only print test results
purrr::walk2(
  list(normal_serie, serie_autocorr),
  c("Normal", "AR(1)"),
  \(serie, name) check_univariate_autocorr(serie, name, test_type = "Box-Pierce",
                                           return_output = FALSE)
)

# Return only the res on tibble format and with lag 1 to 20 tested
df_res <-
  purrr::map2(
    list(normal_serie, serie_autocorr),
    c("Normal", "AR(1)"),
    \(serie, name) check_univariate_autocorr(serie, name, test_type = "Box-Pierce",
                                             return_output = FALSE,
                                             n_lag = 1:20)[["df_res"]]
  ) |>
  purrr::list_rbind() |>
  print()

}
\seealso{
\code{\link[stats]{Box.test}}() For more information on statistical tests.
}
