% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_univariate_normality.R
\name{check_univariate_normality}
\alias{check_univariate_normality}
\title{Check the normality of a univariate serie}
\usage{
check_univariate_normality(
  serie,
  name,
  print_tests = TRUE,
  return_output = TRUE,
  path_output = NULL
)
}
\arguments{
\item{serie}{A numerical vector containing the serie to be tested out.}

\item{name}{Character indicating the name of the serie.}

\item{print_tests}{Logical indicating whether statistical tests should be
printed or not. TRUE by default.}

\item{return_output}{Logical indicating whether the graphs output should be
returned or not. TRUE by default.}

\item{path_output}{NULL or character indicating the path to save the graphs.
If NULL graphs will not be saved. NULL by default.}
}
\value{
Return (if wanted) the normality and qqplot graph in one unique ggplot
object.
}
\description{
Perform statistical tests and draw graphs to test whether a serie
follow a normal distribution or not.
}
\details{
2 type of graphs are drawn :
\itemize{
\item A density graph wich compare the empirical density of the serie to the
normal density where the normal distribution is characterized by the same mean and
standard deviation than the real serie.
\item A QQplot graph wich compare The empirical quantiles distribution to
theoretical quantiles if the serie follow a normal distribution.
}

3 tests statistical tests are performed :
\itemize{
\item Agostino's test : H0 : the serie has a skewness of 0 ; H1 : the serie
has a skewness different of 0.
\item Anscombe's test : H0 : the serie has a kurtosis of 3 ; H1 : the serie has
a kurtosis different of 3
\item Jarque-Berra's tests : H0 : the serie follow a normal distribution ; H1 :
the serie does not follow the normal distribution.
}
}
\examples{
normal_serie <- stats::rnorm(1000, mean = 0, sd = 1)
check_univariate_normality(
  normal_serie,
  "Normal serie"
)

student_serie <- stats::rt(1000, df = 5)
check_univariate_normality(
  student_serie,
  "Student serie"
)

}
\seealso{
\code{\link[moments]{agostino.test}} pour the Agostino's skewness test.
\code{\link[moments]{anscombe.test}} pour the Anscombe's kurtosis test.
\code{\link[moments]{jarque.test}} pour the Jarque-Berra's normality test.
}
