% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parameter.R
\name{.check_data_frame}
\alias{.check_data_frame}
\title{Check if parameter is a data.frame}
\usage{
.check_data_frame(param, name_param)
}
\arguments{
\item{param}{Value of the parameter}

\item{name_param}{Character indicating the name of the parameter}
}
\value{
Stop or nothing
}
\description{
Check if parameter is a data.frame
}
