% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_parameter.R
\name{.check_length}
\alias{.check_length}
\title{Check if Parameter is Length 1}
\usage{
.check_length(param, name_param, ln = 1)
}
\arguments{
\item{param}{Value of the parameter}

\item{name_param}{Character : name of the parameter}

\item{ln}{Numeric indicating the length the parameter must have exactly}
}
\value{
Stop or nothing
}
\description{
Check if Parameter is Length 1
}
