% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granger_coint.R
\name{granger_coint}
\alias{granger_coint}
\title{Perform a cointegration granger test}
\usage{
granger_coint(
  df,
  formula,
  type = c("Moving", "None"),
  window = integer(0),
  nb_min_obs = 520,
  set_seed = TRUE,
  seed = 1234567
)
}
\arguments{
\item{df}{Dataframe cointaing data and a "date" variable.}

\item{formula}{Character or formula indicating the formula of the
long-term regression.}

\item{type}{Character indicating if the granger tests should be condcuted
in a "Moving way" (either reccursive if \code{window = 0} or rolling if
\code{window > 0}) or juste conducted on the full sample. "Moving" by default.}

\item{window}{Numeric indicating the size of the window. 0 for reccursive test.
It must be equal or greater than nb_min_obs.}

\item{nb_min_obs}{Numeric indicating the minimal number of observation the
window must have before the VAR is computed.}

\item{set_seed}{Logical indicating whether a seed should be set to always
obtain same random numbers. TRUE by default. The seed generation is from the
"L'Ecuyer-CMRG" algorithm. See \code{\link[base]{set.seed}}() and
\code{\link[furrr]{furrr_options}}().}

\item{seed}{Numeric indicating the seed number to be used.}
}
\value{
A tibble containing results of rolling and full sample test
(statistical values and critical values).
}
\description{
The test is conducted on the full sample, but can also be conducted on
a rolling window or reccursive way. A long term regression is run with the
formula provided. An ADF test (see \code{\link[dobby]{adf_test_auto}}()) is
then conducted on the residuals. Is there is no unit root, then the series
are said to be cointegrated.
}
\examples{
# No examples yet
}
