% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adf_test_auto.R
\name{model_test}
\alias{model_test}
\title{Test the Unit root for on ADF model}
\usage{
model_test(
  y_name,
  y_diff,
  tt,
  y_lag_1,
  x,
  lags,
  selectlags,
  num_model,
  signif,
  n
)
}
\arguments{
\item{y_name}{String indicating the name of the serie.}

\item{y_diff}{Numeric vector containing the serie in first difference.}

\item{tt}{Numeric vector containing the linear trend.}

\item{y_lag_1}{Numeric vector containing the serie lagged of one period.}

\item{x}{Numeric matrix containing the lagged serie in difference for all
lags.}

\item{lags}{Numeric indicating the maximum number of lags to be tested if
\code{selectlags == "AIC"} or if \code{selectlags == "BIC"}. if \code{selectlags == "Fixed"},
the it is the number of lag to be include.}

\item{selectlags}{String indicating if the number of lags is choose by the
user (\code{selectlags = "Fixed"}) or is choosent by minimizing informations
criterion (\code{selectlags == "AIC"} or \code{selectlags == "BIC"}).}

\item{num_model}{String indicating the model to be tested : "1" for the model
without trend and drift. "2" for the model with drift and without trend.
"3" for the model with drift and trend.}

\item{signif}{Numeric indicating the significance threshold to be used. Must
be one of : 0.01, 0.05 or 0.1.}

\item{n}{Numeric indicating the number of observations}
}
\value{
A list containing the message of the result ; a logical indicating
if the tested parameter (trend or drift) is significative (always TRUE for
model 1) ; A string indicating the number of the model ; The t-stat of the
phi parameter (NA if the trend or drift parameter is not significative) ;
the number of lags choosen.
}
\description{
Perform the ADF test for one of the 3 possible ADF models.
}
\examples{
# Not needed

}
