% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stat.R
\name{summary_stat}
\alias{summary_stat}
\title{Provide a statistical summary for a univariate time series}
\usage{
summary_stat(df, var, serie_name, date_variable = "date")
}
\arguments{
\item{df}{Dataframe containing the variable of interest and the date_variable.}

\item{var}{Character indicating the name of the variable to be studied.}

\item{serie_name}{Character indicating the how to name the serie.}

\item{date_variable}{Character indicating the variable containing the dates
of the serie. By default it is "date".}
}
\value{
A tibble containing the name of the variable ; the beggining and
ending dates the serie ; ADF test results ; Skewness and Kurtosis statistics.
}
\description{
Provide statistical information about a univariate time serie.
Results contained start and end date, ADF test
(see \code{\link{adf_test_auto}}) and the skweness and kurtosis statistics.
}
\examples{
# Create data
data <-
  dplyr::tibble (
    date = seq(from = as.Date("1980-01-01"), by = "day", length.out = 1000),
    normal_serie = stats::rnorm(1000, mean = 0, sd = 1),
    student_serie = stats::rt(1000, df = 5)
  )

# Descriptives statistics for one serie
summary_stat(data, "normal_serie", "Normal")

# Descriptives statistics for 2 series
purrr::map2(
  c("normal_serie", "student_serie"),
  c("Normal", "Student"),
  \(variable, name) summary_stat(data, variable, name)
) |>
  purrr::list_rbind()

}
\seealso{
\itemize{
\item \code{\link[=adf_test_auto]{adf_test_auto()}} For more informations on the ADF test.
\item \code{\link[moments]{skewness}}() For more information on the skewness
statistics
\item \code{\link[moments]{kurtosis}}() For more information on the kurtosis
statistics
}
}
